--- api/src/main/java/org/flab/api/global/exception/ErrorResponse.java
@@ -1,23 +1,0 @@
-package org.flab.api.global.exception;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ErrorResponse {
-
-    private String code;
-    private String message;
-
-    public ErrorResponse(ErrorCode errorCode) {
-        this.code = errorCode.getCode();
-        this.message = errorCode.getMessage();
-    }
-
-    public ErrorResponse(ErrorCode errorCode, String message) {
-        this.code = errorCode.getCode();
-        this.message = message;
-    }
-}
